lesson view report draft

+++ app/src/main/java/org/stepik/android/domain/lesson/model/LessonDeepLinkData.kt
@@ -2,7 +2,7 @@ package org.stepik.android.domain.lesson.model
 
 data class LessonDeepLinkData(
     val lessonId: Long,
-    val stepPosition: Long = 1,
+    val stepPosition: Int = 1,
     val unitId: Long?,
     val discussionId: Long?
 )
